added component for more logic
